requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
camelot-py==0.9.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
distro==1.8.0
et-xmlfile==1.1.0
Flask==2.3.2
html-to-json==2.0.0
html5lib==1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==2.0.1
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.17
PyMuPDF==1.22.3
pypdf==3.8.1
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
tabula-py==2.7.0
tabulate==0.9.0
tzdata==2023.3
Wand==0.6.11
webencodings==0.5.1
Werkzeug==2.3.4
